+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 1.5.1 - 2023-01-18
+### Added
+- Show error (X) if previous command returned nonzero value.
 ### Changed
 - Refactor build system